\name{datasim-package}
\alias{datasim-package}
\alias{datasim}
\docType{package}
\title{
\packageTitle{datasim}
}
\description{
\packageDescription{datasim}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{datasim}
\packageIndices{datasim}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{datasim}

Maintainer: \packageMaintainer{datasim}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
