% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasim-package.R
\docType{package}
\name{datasim}
\alias{datasim}
\alias{datasim-package}
\title{datasim: Data simulation based on models formulas}
\description{
The \code{datasim} package provide tools to simulate data frames based on models
formulas. It works with different types of effects and multivariate models. The
user define a list of formulas than later is used to i) simulate the covariates
presented in the formula and to ii) simulate the response variable.
}
\author{
Erick A. Chacon-Montalvan
}
