% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasim-package.R
\docType{package}
\name{day2day}
\alias{datasim}
\alias{datasim-package}
\alias{day2day-package}
\title{datasim: Data simulation based on models formulas}
\description{
The \code{datasim} package provide tools to simulate data frames based on models
formulas. It works with different types of effects and multivariate models.
}
\author{
Erick A. Chacon-Montalvan
}
