% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effects-uni.R
\name{fa}
\alias{fa}
\title{Univariate factor effects}
\usage{
fa(x, beta, labels = 1:length(beta), size = NULL)
}
\arguments{
\item{x}{factor vector to evaluate the effect}

\item{beta}{numeric vector of the effects for each level of \code{x}}

\item{labels}{character vector to name the \code{levels} of \code{x}}

\item{size}{numeric value to simulate the covariate \code{x}}
}
\value{
return.
}
\description{
\code{function} description.
}
\details{
details.
}
\examples{

(x <- fa(beta = -1:1, size = 10))
fa(x, beta = -1:1)
fa(x, beta = 1:3)

}
\author{
Erick A. Chacon-Montalvan
}
