% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effects-uni.R
\name{gp}
\alias{gp}
\title{Simulate a Gaussian process}
\usage{
gp(..., cor.model = NULL, cor.params = NULL, sigma2 = 1, size = NULL)
}
\arguments{
\item{...}{coordinates}

\item{cor.model}{A character or function indicating the covariance function that
Should be used to compute the variance-covariance matrix}

\item{cor.params}{A list of the parameters required by the \code{cor.model} function.}

\item{sigma2}{variance of the Gaussian process}

\item{size}{numeric value to simulate the covariate \code{x}}
}
\value{
A vector of the realization of the Gaussian Process
}
\description{
\code{gp} Simulate a spatial Gaussian process given a certain covariance function.
}
\details{
details.
}
\examples{

(x <- gp(s1 = NA, size = 10))
(s1 <- x[[1]])
# Simulate and plot the realization of a Gaussian process
(y <- gp(s1, cor.model = "exp_cor", cor.params = list(phi = 0.05)))

(x <- gp(s1 = NA, s2 = NA, size = 10))
(s1 <- x[[1]])
(s2 <- x[[2]])
# Simulate and plot the realization of a Gaussian process
(y <- gp(s1, s2, cor.model = "exp_cor", cor.params = list(phi = 0.05)))
plot(s1, s2, cex = y)
# Plot with ggplot
# ggplot(data.frame(s1, s2, y), aes(s1, s2, col = y)) +
#  geom_point(size = 3)

}
\author{
Erick A. Chacon-Montalvan
}
