% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effects-multi.R
\name{mfe}
\alias{mfe}
\title{Multivariate Fixed Effect}
\usage{
mfe(x, beta)
}
\arguments{
\item{x}{A vector of length \code{n} for which the fixed effect will be evaluated.}

\item{beta}{A vector of length \code{q}, this is the fixed effect for each response
variable.}
}
\value{
A matrix of dimension n x q.
}
\description{
\code{mfe} compute the multivariate fixed effect. Generally used with
\code{sim_model}.
}
\details{
details.
}
\examples{

mfe(x = rep(rnorm(5), 3), beta = c(0.1, 0, 1))

mfe(x = rep(rnorm(5), 3), beta = rep(10, 3))

}
\author{
Erick A. Chacon-Montalvan
}
