% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effects-multi.R
\name{mgp}
\alias{mgp}
\title{Simulate a Multivariate Gaussian process}
\usage{
mgp(..., variance = NULL, cor.model = NULL, cor.params = NULL,
  size = NULL)
}
\arguments{
\item{...}{coordinates}

\item{variance}{A qxq cross-covariance matrix.}

\item{cor.model}{A character or function indicating the covariance function that
should be used to compute the correlation matrix}

\item{cor.params}{List of lists indicating the parameters for each response.}

\item{size}{numeric value to simulate the covariate \code{x}}
}
\value{
A vector of the realization of the Gaussian Process
}
\description{
\code{mgp} Simulate a Multivariate spatial Gaussian process known as linear model
of coregionalization Y(h) = AS(h), where S(h) is a vector of q independent Gaussian
processes.
}
\details{
details.
}
\examples{

# Generate coordinates
N <- 100
s1 <- 2 * runif(N)
s2 <- 2 * runif(N)

# Covariance parameters
q <- 2
var <- sqrt(diag(c(4, 4)))
A <- matrix(c(1, - 0.8, 0, 0.6), nrow = 2)
variance <- var \%*\% tcrossprod(A) \%*\% var
cor.params <- list(list(phi = 0.08), list(phi = 0.08))

# Generate the multivariate Gaussian process
y <- mgp(s1, s2, variance = variance, cor.model = "exp_cor", cor.params =
cor.params)
y1 <- y[1:N]
y2 <- y[(N + 1):(2 * N)]

# Check correlation
cor(y1, y2)
plot(y1, y2)

# Visualize the spatial
plot(s1, s2, cex = y1, col = 2)
points(s1, s2, cex = y2, col = 3)

(x <- mgp(s1 = NA, s2 = NA, size = 100))
(s1 <- x[[1]])
(s2 <- x[[2]])
cor.model <- "exp_cor"
cor.params <- list(list(phi = 0.05), list(phi = 0.07))
variance = matrix(c(2, 1.5, 1.5, 2), nrow = 2)
(out <- mgp(s1, s2, variance = variance, cor.model = cor.model, cor.params = cor.params))
out_mat <- matrix(out, ncol = 2)
plot(out_mat)
cov(out_mat)

(x <- mgp(s1 = NA, size = 10))
(s1 <- x[[1]])
cor.model <- "exp_cor"
cor.params <- list(list(phi = 0.05), list(phi = 0.07), list(phi = 1))
variance <- matrix(c(1, 0.8, 0.5, 0.8, 1, 0.5, 0.5, 0.5, 1), nrow = 3)
mgp(s1, variance = variance, cor.model = cor.model, cor.params = cor.params)

(x <- mgp(s1 = NA, size = 100))
(s1 <- x[[1]])
cor.model <- "exp_cor"
cor.params <- list(list(phi = 0.05), list(phi = 0.07), list(phi = 1))
variance <- matrix(c(1, 0.8, 0.5, 0.8, 1, 0.5, 0.5, 0.5, 1), nrow = 3)
mgp(s1, variance = variance, cor.model = cor.model, cor.params = cor.params)

}
\author{
Erick A. Chacon-Montalvan
}
