% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim-model.R
\name{model_frame}
\alias{model_frame}
\title{Simulate covariates based on model formula}
\usage{
model_frame(formula, n = nrow(idata), idata = NULL, seed = NULL)
}
\arguments{
\item{formula}{List of the parameters, indicating how they should be computed.
similar to formula for \code{lm}, \code{glm}, \code{bamlss}, with the difference
that it included the coefficients and link function explicitly.}

\item{n}{Number of observations to be simulated}

\item{idata}{Initial data including some variables to not been simulated.}

\item{seed}{Seed to be defined with function \code{set.seed} to obtain reproducible}
}
\value{
return.
}
\description{
\code{function} description.
}
\details{
details.
}
\examples{

# Structure of the model
formula <- list(
  mean ~ I(age ^ 2) + fa(sex, beta = c(-1, 1)),
  sd ~ fa(sex, beta = c(1, 2))
)
idata <- data.frame(s1 = 1:10)
(datasim <- model_frame(formula, idata = idata))
(datasim <- model_frame(formula, n = 10))

formula <- list(
  mean ~ I(5) + I(0.5 * x1) + I(0.1 * x2),
  sd ~ I(2) + I(x)
)

# Structure of the model
formula <- list(
  mean ~ fa(sex, beta = c(-1,1)),
  sd ~ I(0)
)

}
\author{
Erick A. Chacon-Montalvan
}
