% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim-model.R
\name{model_response_lm}
\alias{model_response_lm}
\title{Simulate response variable (testing function)}
\usage{
model_response_lm(model_frame, formula = attr(model_frame, "formula"),
  link_inv = replicate(length(formula), identity), generator = rnorm,
  responses = c("response"), effects_save = FALSE, seed = NULL)
}
\arguments{
\item{model_frame}{A \code{tibble} containing all the covariates to be used.
Usually this is obtained as the output of the \code{model_frame} function.}

\item{formula}{An optional list of formulas to simulate the parameters of the
response variables.}

\item{link_inv}{A list of function representing the inverse link function for the
parameters.}

\item{generator}{Function to generate the response variables given the parameters}

\item{responses}{character vector indicating the names of the response variables}

\item{effects_save}{Optional logical argument to save or not generated random
effects}

\item{seed}{Seed to be defined with function \code{set.seed} to obtain reproducible
results}
}
\description{
\code{function} description.
}
\examples{

formula <- list(
mean ~ I(0.5 * x1) : I(x2) + re(city, 1, 2),
sd ~ I(1)
)
model_frame <- model_frame(formula, n = 10)
model_response_lm(model_frame)



}
