% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effects-multi.R
\name{st_sample}
\alias{st_sample}
\title{bla}
\usage{
st_sample(..., size)
}
\arguments{
\item{par.}{}
}
\value{
return.
}
\description{
\code{function} description.
}
\details{
details.
}
\examples{



}
\author{
Erick A. Chacon-Montalvan
}
