% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effects-uni.R
\name{gp}
\alias{gp}
\title{Simulate a Gaussian process}
\usage{
gp(coords, cor.model = "exp_cor", cor.params, sigma2 = 1,
  size = NULL, range = 1, geom = NULL)
}
\arguments{
\item{coords}{A list of coordinates}

\item{cor.model}{A character or function indicating the covariance function that
Should be used to compute the variance-covariance matrix}

\item{cor.params}{A list of the parameters required by the \code{cor.model} function.}

\item{sigma2}{variance of the Gaussian process}

\item{size}{numeric value to simulate the covariate \code{x}}

\item{range}{Range of the coordinates of the Gaussian processes.}

\item{geom}{object of class ‘sf’ or ‘sfc’ to define the area where to simulate the
samples}
}
\value{
A vector of the realization of the Gaussian Process
}
\description{
\code{gp} Simulate a spatial Gaussian process given a certain covariance function.
}
\details{
details.
}
\examples{

(x <- gp(list(s1 = NA), size = 10))
(s1 <- x[[1]])
# Simulate and plot the realization of a Gaussian process
(y <- gp(list(s1), cor.model = "exp_cor", cor.params = list(phi = 0.05)))

(x <- gp(list(s1 = NA, s2 = NA), size = 10))
(s1 <- x[[1]])
(s2 <- x[[2]])
# Simulate and plot the realization of a Gaussian process
(y <- gp(list(s1, s2), cor.model = "exp_cor", cor.params = list(phi = 0.05)))
plot(s1, s2, cex = y)
# Plot with ggplot
# ggplot(data.frame(s1, s2, y), aes(s1, s2, col = y)) +
#  geom_point(size = 3)

(x <- gp(list(s1 = "none", s2 = NA), size = 10))
(s1 <- x[[1]])
(s2 <- s1)
# Simulate and plot the realization of a Gaussian process
(y <- gp(list(s1, s2), cor.model = "exp_cor", cor.params = list(phi = 0.05)))

}
\author{
Erick A. Chacón-Montalván
}
