% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effects-multi.R
\name{mfa}
\alias{mfa}
\title{Multivariate factor effects}
\usage{
mfa(x, beta, labels = 1:nrow(beta), size = NULL)
}
\arguments{
\item{x}{factor vector to evaluate the effect}

\item{beta}{matrix of the effects for each level of \code{x}}

\item{labels}{character vector to name the \code{levels} of \code{x}}

\item{size}{numeric value to simulate the covariate \code{x}}
}
\value{
return.
}
\description{
\code{function} description.
}
\details{
\code{mfa} has two behaviours. If \code{size == NULL}, it evaluate the effects
applied to the argument \code{x}; otherwise, it generates a factor \code{x}.
}
\examples{

(x <- mfa(beta = cbind(1:2, 2:3, 0:1), size = 10))
mfa(x, beta = cbind(1:2, 2:3, 0:1))

(x <- mfa(beta = replicate(3, 0:2), size = 10))
mfa(x, beta = replicate(3, 0:2))

(x <- mfa(beta = cbind(1, 2, 0), size = 10))
mfa(x, beta = cbind(1, 2, 0))

}
\author{
Erick A. Chacon-Montalvan
}
