% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effects-multi.R
\name{mgp}
\alias{mgp}
\title{Simulate a Multivariate Gaussian process}
\usage{
mgp(coords, variance, cor.model, cor.params, size = NULL, range = 1)
}
\arguments{
\item{coords}{A list of coordinates.}

\item{variance}{A qxq cross-covariance matrix Sigma such as \code{A=chol(Sigma)}.}

\item{cor.model}{A character or function indicating the covariance function that
should be used to compute the correlation matrix}

\item{cor.params}{List of lists indicating the parameters for each response.}

\item{size}{numeric value to simulate the covariate \code{x}}

\item{range}{range on the coordinates of the Gaussian process}
}
\value{
A vector of the realization of the Gaussian Process
}
\description{
\code{mgp} Simulate a Multivariate spatial Gaussian process known as linear model
of coregionalization Y(h) = AS(h), where S(h) is a vector of q independent Gaussian
processes.
}
\details{
details.
}
\examples{

# Generate coordinates
N <- 100
s1 <- rep(2 * runif(N), 2)
s2 <- rep(2 * runif(N), 2)

# Covariance parameters
q <- 2
var <- sqrt(diag(c(4, 4)))
A <- matrix(c(1, - 0.8, 0, 0.6), nrow = 2)
variance <- var \%*\% tcrossprod(A) \%*\% var
cor.params <- list(list(phi = 0.08), list(phi = 0.08))

# Generate the multivariate Gaussian process
y <- mgp(list(s1, s2), variance, "exp_cor", cor.params)
y1 <- y[1:N]
y2 <- y[(N + 1):(2 * N)]

# Check correlation
cor(y1, y2)
plot(y1, y2)

# Visualize the spatial
plot(s1, s2, cex = y1, col = 2)
points(s1, s2, cex = y2, col = 3)

variance = matrix(c(2, 1.5, 1.5, 2), nrow = 2)
(x <- mgp(list(s1 = NA, s2 = NA), variance, size = 100))
(s1 <- x[[1]])
(s2 <- x[[2]])
cor.model <- "exp_cor"
cor.params <- list(list(phi = 0.05), list(phi = 0.07))
(out <- mgp(list(s1, s2), variance, cor.model, cor.params))
out_mat <- matrix(out, ncol = 2)
plot(out_mat)
cov(out_mat)

variance <- matrix(c(1, 0.8, 0.5, 0.8, 1, 0.5, 0.5, 0.5, 1), nrow = 3)
(x <- mgp(list(s1 = NA), variance, size = 10))
(s1 <- x[[1]])
cor.model <- "exp_cor"
cor.params <- list(list(phi = 0.05), list(phi = 0.07), list(phi = 1))
mgp(list(s1), variance = variance, cor.model = cor.model, cor.params = cor.params)

variance <- matrix(c(1, 0.8, 0.5, 0.8, 1, 0.5, 0.5, 0.5, 1), nrow = 3)
(x <- mgp(list(s1 = NA), variance, size = 100))
(s1 <- x[[1]])
cor.model <- "exp_cor"
cor.params <- list(list(phi = 0.05), list(phi = 0.07), list(phi = 1))
mgp(list(s1), variance, cor.model, cor.params)

}
\author{
Erick A. Chacón-Montalván
}
