% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effects-multi.R
\name{mgp}
\alias{mgp}
\title{Multivariate Gaussian Processes}
\usage{
mgp(coords, A, cor.model, cor.params, size = NULL, range = 1)
}
\arguments{
\item{coords}{A list of replicated coordinates. Each coordinate is a \eqn{q} replicated
vector of the unique coordinates. If \code{size != NULL}, \code{coords} is the
output of the function \code{mgp}.}

\item{A}{A qxm matrix that defines the relationship between \eqn{Y(h)} and
\eqn{S(h)}.}

\item{cor.model}{A character or function indicating the correlation function to
be used to compute the correlation matrix}

\item{cor.params}{A nested list of the parameters of the correlation function for
each response.}

\item{size}{A numeric value \eqn{n} representing the number of locations, it is used to
simulate the coordinates \code{coords}. In case \code{size == NULL}, \code{mgp}
simulates the multivariate Gaussian process.}

\item{range}{Range of the coordinates of the Gaussian processes.}
}
\value{
A list of simulated replicated coordinates in case \code{size} is provided;
otherwise, a \eqn{nq}-length numeric vector of the evaluated multivariate Gaussian
process.
}
\description{
\code{mgp} handles the evaluation of multivariate Gaussian processes with two
behaviours. It evaluates the multivariate effects applied to the coordinates
\code{coords} if \code{size == NULL} or it simulates the coordinates
\code{coords} if \code{size} is provided.
}
\details{
The \eqn{q}-dimensional multivariate Gaussian process \eqn{Y(h)} is
expressed as a linear combination
\deqn{Y(h) = AS(h).}
where \eqn{S(h)} is a vector of \eqn{m} standardized independent Gaussian
processes at location \eqn{h} and \eqn{A} a \eqn{qxm} matrix of coefficients.
}
\examples{
## Simulation of a temporal multivariate Gaussian process

# Define the number of location and responses, and a linear transformation matrix
n <- 200
q <- 3
A <- matrix(c(1, 0, 1, 0, 1, - 0.8), nrow = q)

# Simulate coordinates
coords <- mgp(list(time = NA), A, size = n)

# Simulate a multivariate Gaussian process
cor.params <- list(list(phi = 0.08), list(phi = 0.15))
y <- mgp(coords, A, "exp_cor", cor.params)

# Visualize the temporal multivariate Gaussian process
data <- data.frame(time = coords[[1]], y, response = factor(rep(1:q, each = n)))
library(ggplot2)
ggplot(data, aes(time)) +
  geom_line(aes(y = y, col = response))

## Simulation of a spatial multivariate Gaussian process

# Define the number of location and responses, and a linear transformation matrix
n <- 1500
q <- 2
A <- matrix(c(1, -1), nrow = q)

# Simulate coordinates
coords <- mgp(list(s1 = NA, s2 = NA), A, size = n)

# Simulate a multivariate Gaussian process
cor.params <- list(list(phi = 0.1))
y <- mgp(coords, A, "exp_cor", cor.params)

# Visualize the temporal multivariate Gaussian process
data <- data.frame(s1 = coords[[1]], s2 = coords[[2]], y,
  response = factor(rep(1:q, each = n)))
ggplot(data, aes(s1, s2)) +
  geom_point(aes(size = y, col = y)) +
  facet_wrap(~ response)

}
\author{
Erick A. Chacón-Montalván
}
