% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effects-multi.R
\name{mre}
\alias{mre}
\title{Multivariate random effects}
\usage{
mre(x, sigma, groups, size = NULL, replace = TRUE)
}
\arguments{
\item{x}{factor vector to evaluate the random effect}

\item{sigma}{variance matrix of the random effect}

\item{groups}{character vector to name the \code{levels} of \code{x} or a numeric
value indicating the number of groups of \code{x}}

\item{size}{numeric value to simulate the covariate \code{x}}

\item{replace}{An optional argument to simulate an independent random effect with
one repetition (e.g. \code{mre(groups = 100, size = 100, replace = FALSE)})}
}
\value{
return.
}
\description{
\code{function} description.
}
\details{
details.
}
\examples{

Sigma <- matrix(c(1, 0.8, 0.5, 0.8, 1, 0.5, 0.5, 0.5, 1), nrow = 3)
(x <- mre(groups = 300, sigma = Sigma, size = 300))
(effect <- mre(x, sigma = Sigma))

id_uni <- match(unique(x), x)
effect_mat <- matrix(effect, ncol = nrow(Sigma))[id_uni, ]
cov(effect_mat)

(x <- mre(groups = 300, sigma = matrix(c(2, 1.5, 1.5, 2), nrow = 2), size = 3000))
(effect <- mre(x, sigma = matrix(c(2, 1.5, 1.5, 2), nrow = 2)))

id_uni <- match(unique(x), x)
effect_mat <- matrix(effect, ncol = 2)[id_uni, ]
cov(effect_mat)

(x <- mre(groups = 10, sigma = matrix(c(2, 1.5, 1.5, 2), nrow = 2),
  size = 10, replace = FALSE))
(effect <- mre(x, sigma = matrix(c(2, 1.5, 1.5, 2), nrow = 2)))

id_uni <- match(unique(x), x)
effect_mat <- matrix(effect, ncol = 2)[id_uni, ]
cov(effect_mat)

}
\author{
Erick A. Chacon-Montalvan
}
